+++ packages/rum-core/src/domain/action/actionCollection.ts
@@ -7,8 +7,6 @@ import type { LifeCycle, RawRumEventCollectedData } from '../lifeCycle'
 import { LifeCycleEventType } from '../lifeCycle'
 import type { RumConfiguration } from '../configuration'
 import type { CommonContext } from '../contexts/commonContext'
-import type { PageStateHistory } from '../contexts/pageStateHistory'
-import { PageState } from '../contexts/pageStateHistory'
 import type { RumActionEventDomainContext } from '../../domainContext.types'
 import type { PartialRumEvent, Hooks } from '../../hooks'
 import { HookNames } from '../../hooks'
@@ -32,11 +30,10 @@ export function startActionCollection(
   hooks: Hooks,
   domMutationObservable: Observable<void>,
   windowOpenObservable: Observable<void>,
-  configuration: RumConfiguration,
-  pageStateHistory: PageStateHistory
+  configuration: RumConfiguration
 ) {
   lifeCycle.subscribe(LifeCycleEventType.AUTO_ACTION_COMPLETED, (action) =>
-    lifeCycle.notify(LifeCycleEventType.RAW_RUM_EVENT_COLLECTED, processAction(action, pageStateHistory))
+    lifeCycle.notify(LifeCycleEventType.RAW_RUM_EVENT_COLLECTED, processAction(action))
   )
 
   hooks.register(HookNames.Assemble, ({ startTime, eventType }): PartialRumEvent | undefined => {
@@ -75,7 +72,7 @@ export function startActionCollection(
     addAction: (action: CustomAction, savedCommonContext?: CommonContext) => {
       lifeCycle.notify(LifeCycleEventType.RAW_RUM_EVENT_COLLECTED, {
         savedCommonContext,
-        ...processAction(action, pageStateHistory),
+        ...processAction(action),
       })
     },
     actionContexts,
@@ -83,10 +80,7 @@ export function startActionCollection(
   }
 }
 
-function processAction(
-  action: AutoAction | CustomAction,
-  pageStateHistory: PageStateHistory
-): RawRumEventCollectedData<RawRumActionEvent> {
+function processAction(action: AutoAction | CustomAction): RawRumEventCollectedData<RawRumActionEvent> {
   const autoActionProperties = isAutoAction(action)
     ? {
         action: {
@@ -114,32 +108,25 @@ function processAction(
         },
       }
     : undefined
-  const customerContext = !isAutoAction(action) ? action.context : undefined
   const actionEvent: RawRumActionEvent = combine(
     {
-      action: {
-        id: generateUUID(),
-        target: {
-          name: action.name,
-        },
-        type: action.type,
-      },
+      action: { id: generateUUID(), target: { name: action.name }, type: action.type },
       date: action.startClocks.timeStamp,
       type: RumEventType.ACTION as const,
-      view: { in_foreground: pageStateHistory.wasInPageStateAt(PageState.ACTIVE, action.startClocks.relative) },
     },
     autoActionProperties
   )
 
-  const domainContext: RumActionEventDomainContext = isAutoAction(action) ? { events: action.events } : {}
-
-  if (!isAutoAction(action) && action.handlingStack) {
-    domainContext.handlingStack = action.handlingStack
-  }
+  const duration = isAutoAction(action) ? action.duration : undefined
+  const customerContext = !isAutoAction(action) ? action.context : undefined
+  const domainContext: RumActionEventDomainContext = isAutoAction(action)
+    ? { events: action.events }
+    : { handlingStack: action.handlingStack }
 
   return {
     customerContext,
     rawRumEvent: actionEvent,
+    duration,
     startTime: action.startClocks.relative,
     domainContext,
   }